Fix add vault (đừng lỗi nữa please), scrollview cho edit bar

+++ app/src/main/java/com/team2/higallery/utils/DataUtils.java
@@ -117,7 +117,7 @@ public class DataUtils {
     // Loại bỏ ảnh trùng lặp, trả về số lượng ảnh bị loại bỏ
     public static int deleteDuplicateImages(Context context) {
         int count = 0;
-        HashMap<Integer, ArrayList<Bitmap>> hashMap = new HashMap<>();
+        HashMap<Long, ArrayList<Bitmap>> hashMap = new HashMap<>();
         TrashManager trashManager = TrashManager.getInstance(context);
 
         // Duyệt tất cả các path ảnh
@@ -129,7 +129,7 @@ public class DataUtils {
             Bitmap currBitmap = BitmapFactory.decodeFile(imagePath);
 
             // Lấy mã hash của bitmap
-            int hashCode = BitmapUtils.getHashCode(currBitmap);
+            long hashCode = BitmapUtils.getHashCode(currBitmap);
 
             // Nếu trong map đã có hashCode này
             if (hashMap.containsKey(hashCode)) {
